reorganize server files into module
