+++ siliconcompiler/remote/server_schema/delete_job.json
